Fix memory ownership and undefined behaviour across MARS
Repair container alignment, allocation overflow checks, numerical special-function edge cases, and expression/differential-equation ownership. Initialise expression singletons on first evaluation or rendering.

Add regressions, separate large matrix and wave test groups, synchronise ownership documentation, and record completed sequential sanitizer and Valgrind verification.

+++ src/expression/expr_maths_ops.c
@@ -414,7 +414,7 @@ static number_t eval_finite_summation(expr_t *dv)
                 return weighted_value;
             num_destroy(&weighted_value);
         }
-        if ((unsigned long)upper_value - (unsigned long)lower_value >= maximum_streamed_terms)
+        if ((unsigned long)upper_value - (unsigned long)lower_value >= (unsigned long)maximum_streamed_terms)
             return num_clone(NUM_NAN);
     }
